Gallery Page changes
